BusyReader is now working correctly (fixed initiation)
