+++ DoodleServer/src/ch/unibe/scg/doodle/DooMockup.java
@@ -1,10 +1,6 @@
 package ch.unibe.scg.doodle;
 
 import ch.unibe.scg.doodle.helperClasses.DoodleList;
-import ch.unibe.scg.doodle.inject.DoodleModule;
-
-import com.google.inject.Guice;
-import com.google.inject.Injector;
 
 /**
  * Mockup on server side for <code>Doo</code> on client side.
@@ -13,19 +9,6 @@ import com.google.inject.Injector;
  * 
  */
 public class DooMockup {
-
-	static Injector injector;
-
-	static Injector injectorInstance() {
-		if (injector == null) {
-			injector = Guice.createInjector(new DoodleModule());
-		}
-		return injector;
-	}
-
-	public static void resetInjector() {
-		injector = null;
-	}
 
 	/**
 	 * Code sugar for Doo.dle-ing any Object similar to System.out.println
@@ -45,7 +28,7 @@ public class DooMockup {
 	 *            o
 	 */
 	public static void dle(Object o, Object... objects) {
-		Doodler doodler = injectorInstance().getInstance(Doodler.class);
+		Doodler doodler = Doodler.instance();
 
 		if (objects.length == 0) {
 			doodler.visualize(o);
